+++ src/Services/ProductService.php
@@ -6,6 +6,7 @@ use Ordercloud\Entities\Products\ProductTag;
 use Ordercloud\Requests\Products\Criteria\ProductCriteria;
 use Ordercloud\Requests\Products\FindProductsRequest;
 use Ordercloud\Requests\Products\GetProductRequest;
+use Ordercloud\Requests\Products\GetProductTagRequest;
 use Ordercloud\Requests\Products\GetProductTagsForOrganisationByTypeNameRequest;
 
 class ProductService extends OrdercloudService
@@ -46,6 +47,13 @@ class ProductService extends OrdercloudService
     {
         return $this->request(
             new GetProductTagsForOrganisationByTypeNameRequest($organisationId, $tagName)
+        );
+    }
+
+    public function getProductTag($tagId)
+    {
+        return $this->request(
+            new GetProductTagRequest($tagId)
         );
     }
 }